+++ ChessAndroid/app/src/main/java/net/theluckycoder/chess/Piece.kt
@@ -2,9 +2,9 @@ package net.theluckycoder.chess
 
 class Piece(
     @JvmField
-    val x: Short,
+    val x: Byte,
     @JvmField
-    val y: Short,
+    val y: Byte,
     @JvmField
     val type: Byte
 )
